attempting temporal mean interpolation
the goal is that, for cells in the feature set where there is missing
data, find the mean value of this attribute for this sample point (in
this case, identified by 'id' for simplicity) across all the years in
the overall date range
eg. a data point that is missing its value for ndvi_mean in 2021 will
have that value supplanted with the mean of nvdi_mean recorded at that
point in 2020, 2022, 2023 and 2024.